+++ src/elements/Entry.php
@@ -847,7 +847,7 @@ class Entry extends Element
                 $currentSites = static::find()
                     ->anyStatus()
                     ->id($this->id)
-                    ->siteId('*')
+                    ->site('*')
                     ->select('elements_sites.siteId')
                     ->drafts(null)
                     ->provisionalDrafts(null)
@@ -862,7 +862,7 @@ class Entry extends Element
                 array_push($currentSites, ...static::find()
                     ->anyStatus()
                     ->id($this->duplicateOf->id)
-                    ->siteId('*')
+                    ->site('*')
                     ->select('elements_sites.siteId')
                     ->drafts(null)
                     ->provisionalDrafts(null)
@@ -1178,6 +1178,7 @@ class Entry extends Element
     public function setAuthor(User $author = null)
     {
         $this->_author = $author;
+        $this->authorId = $author->id;
     }
 
     /**
@@ -1541,13 +1542,13 @@ EOD;
         if ($this->_shouldSaveRevision()) {
             $hasRevisions = self::find()
                 ->revisionOf($this)
-                ->siteId('*')
+                ->site('*')
                 ->anyStatus()
                 ->exists();
             if (!$hasRevisions) {
                 $currentEntry = self::find()
                     ->id($this->id)
-                    ->siteId('*')
+                    ->site('*')
                     ->anyStatus()
                     ->one();
 
@@ -1638,20 +1639,7 @@ EOD;
             if (!$this->duplicateOf && $section->type == Section::TYPE_STRUCTURE) {
                 // Has the parent changed?
                 if ($this->_hasNewParent()) {
-                    $mode = $isNew ? Structures::MODE_INSERT : Structures::MODE_AUTO;
-                    if (!$this->newParentId) {
-                        if ($section->defaultPlacement === Section::DEFAULT_PLACEMENT_BEGINNING) {
-                            Craft::$app->getStructures()->prependToRoot($this->structureId, $this, $mode);
-                        } else {
-                            Craft::$app->getStructures()->appendToRoot($this->structureId, $this, $mode);
-                        }
-                    } else {
-                        if ($section->defaultPlacement === Section::DEFAULT_PLACEMENT_BEGINNING) {
-                            Craft::$app->getStructures()->prepend($this->structureId, $this, $this->getParent(), $mode);
-                        } else {
-                            Craft::$app->getStructures()->append($this->structureId, $this, $this->getParent(), $mode);
-                        }
-                    }
+                    $this->_placeInStructure($isNew, $section);
                 }
 
                 // Update the entry's descendants, who may be using this entry's URI in their own URIs
@@ -1664,6 +1652,40 @@ EOD;
         }
 
         parent::afterSave($isNew);
+    }
+
+    private function _placeInStructure(bool $isNew, Section $section): void
+    {
+        // If this is a provisional draft and its new parent matches the canonical entry’s, just drop it from the structure
+        if ($this->isProvisionalDraft) {
+            $canonicalParentId = Entry::find()
+                ->select(['elements.id'])
+                ->ancestorOf($this->getCanonicalId())
+                ->ancestorDist(1)
+                ->anyStatus()
+                ->scalar();
+
+            if ($this->newParentId == $canonicalParentId) {
+                Craft::$app->getStructures()->remove($this->structureId, $this);
+                return;
+            }
+        }
+
+        $mode = $isNew ? Structures::MODE_INSERT : Structures::MODE_AUTO;
+
+        if (!$this->newParentId) {
+            if ($section->defaultPlacement === Section::DEFAULT_PLACEMENT_BEGINNING) {
+                Craft::$app->getStructures()->prependToRoot($this->structureId, $this, $mode);
+            } else {
+                Craft::$app->getStructures()->appendToRoot($this->structureId, $this, $mode);
+            }
+        } else {
+            if ($section->defaultPlacement === Section::DEFAULT_PLACEMENT_BEGINNING) {
+                Craft::$app->getStructures()->prepend($this->structureId, $this, $this->getParent(), $mode);
+            } else {
+                Craft::$app->getStructures()->append($this->structureId, $this, $this->getParent(), $mode);
+            }
+        }
     }
 
     /**
@@ -1751,7 +1773,7 @@ EOD;
                 $drafts = static::find()
                     ->draftOf($this)
                     ->anyStatus()
-                    ->siteId('*')
+                    ->site('*')
                     ->unique()
                     ->all();
                 $structuresService = Craft::$app->getStructures();
@@ -1791,13 +1813,16 @@ EOD;
      */
     private function _checkForNewParent(): bool
     {
-        // Make sure this is a Structure section
-        if ($this->getSection()->type != Section::TYPE_STRUCTURE) {
+        // Make sure this is a Structure section, and that it’s either a canonical entry or provisional drafT
+        if (!(
+            $this->getSection()->type === Section::TYPE_STRUCTURE &&
+            ($this->getIsCanonical() || $this->isProvisionalDraft)
+        )) {
             return false;
         }
 
-        // Is it a brand new entry?
-        if ($this->id === null) {
+        // Is it a brand new (non-provisional) entry?
+        if ($this->id === null && !$this->isProvisionalDraft) {
             return true;
         }
 
@@ -1806,24 +1831,31 @@ EOD;
             return false;
         }
 
+        // If this is a provisional draft, but doesn't actually exist in the structure yet, check based on the canonical entry
+        if ($this->isProvisionalDraft && !$this->lft) {
+            $entry = $this->getCanonical(true);
+        } else {
+            $entry = $this;
+        }
+
         // Is it set to the top level now, but it hadn't been before?
-        if (!$this->newParentId && $this->level != 1) {
+        if (!$this->newParentId && $entry->level != 1) {
             return true;
         }
 
         // Is it set to be under a parent now, but didn't have one before?
-        if ($this->newParentId && $this->level == 1) {
+        if ($this->newParentId && $entry->level == 1) {
             return true;
         }
 
         // Is the parentId set to a different entry ID than its previous parent?
-        $oldParentQuery = self::find();
-        $oldParentQuery->ancestorOf($this);
-        $oldParentQuery->ancestorDist(1);
-        $oldParentQuery->siteId($this->siteId);
-        $oldParentQuery->anyStatus();
-        $oldParentQuery->select('elements.id');
-        $oldParentId = $oldParentQuery->scalar();
+        $oldParentId = self::find()
+            ->ancestorOf($entry)
+            ->ancestorDist(1)
+            ->siteId($entry->siteId)
+            ->anyStatus()
+            ->select('elements.id')
+            ->scalar();
 
         return $this->newParentId != $oldParentId;
     }