+++ src/elements/db/ElementQuery.php
@@ -42,7 +42,7 @@ use yii\db\ExpressionInterface;
 /**
  * ElementQuery represents a SELECT SQL statement for elements in a way that is independent of DBMS.
  *
- * @property string|Site $site The site or site handle that the elements should be returned in
+ * @property-write string|string[]|Site $site The site(s) that resulting elements must be returned in
  * @mixin CustomFieldBehavior
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.0.0
@@ -2433,7 +2433,8 @@ class ElementQuery extends Query implements ElementQueryInterface
     {
         return (
             !$this->trashed &&
-            ($this->withStructure || ($this->withStructure !== false && $this->structureId))
+            !$this->revisions &&
+            ($this->withStructure ?? (bool)$this->structureId)
         );
     }
 
@@ -2476,12 +2477,12 @@ class ElementQuery extends Query implements ElementQueryInterface
             ]);
 
         if ($this->structureId) {
-            $this->query->innerJoin(['structureelements' => Table::STRUCTUREELEMENTS], [
+            $this->query->leftJoin(['structureelements' => Table::STRUCTUREELEMENTS], [
                 'and',
                 '[[structureelements.elementId]] = [[subquery.elementsId]]',
                 ['structureelements.structureId' => $this->structureId],
             ]);
-            $this->subQuery->innerJoin(['structureelements' => Table::STRUCTUREELEMENTS], [
+            $this->subQuery->leftJoin(['structureelements' => Table::STRUCTUREELEMENTS], [
                 'and',
                 '[[structureelements.elementId]] = [[elements.id]]',
                 ['structureelements.structureId' => $this->structureId],
@@ -2645,13 +2646,9 @@ class ElementQuery extends Query implements ElementQueryInterface
         $useCanonicalId = $db->columnExists(Table::ELEMENTS, 'canonicalId');
 
         if ($this->drafts !== false) {
-            if ($this->drafts === true) {
-                $this->subQuery->innerJoin(['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
-                $this->query->innerJoin(['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
-            } else {
-                $this->subQuery->leftJoin(['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
-                $this->query->leftJoin(['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
-            }
+            $joinType = $this->drafts === true ? 'INNER JOIN' : 'LEFT JOIN';
+            $this->subQuery->join($joinType, ['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
+            $this->query->join($joinType, ['drafts' => Table::DRAFTS], '[[drafts.id]] = [[elements.draftId]]');
 
             $this->query->addSelect([
                 'elements.draftId',
@@ -2696,7 +2693,6 @@ class ElementQuery extends Query implements ElementQueryInterface
                     ['drafts.provisional' => $this->provisionalDrafts],
                 ]);
             }
-
 
             if ($this->savedDraftsOnly) {
                 // todo: remove this check after the next breakpoint
@@ -2775,24 +2771,43 @@ class ElementQuery extends Query implements ElementQueryInterface
      */
     private function _normalizeStructureParamValue(string $property, string $class): ElementInterface
     {
-        if ($this->$property !== false && !$this->$property instanceof ElementInterface) {
-            $this->$property = $class::find()
-                ->id($this->$property)
+        $element = $this->$property;
+
+        if ($element === false) {
+            throw new QueryAbortedException();
+        }
+
+        /** @var string|ElementInterface $class */
+        if ($element instanceof ElementInterface && !$element->lft) {
+            $element = $element->id;
+        }
+
+        if (!$element instanceof ElementInterface) {
+            $element = $class::find()
+                ->id($element)
                 ->siteId($this->siteId)
                 ->structureId($this->structureId)
                 ->anyStatus()
                 ->one();
 
-            if ($this->$property === null) {
+            if ($element === null) {
                 $this->$property = false;
+                throw new QueryAbortedException();
             }
         }
 
-        if ($this->$property === false) {
-            throw new QueryAbortedException();
+        if (!$element->lft) {
+            if ($element->getIsDerivative()) {
+                $element = $element->getCanonical(true);
+            }
+
+            if (!$element->lft) {
+                $this->$property = false;
+                throw new QueryAbortedException();
+            }
         }
 
-        return $this->$property;
+        return $this->$property = $element;
     }
 
     /**